Corrige lo que marcó la revisión
Elegir una pantalla en la barra esperaba las preferencias tanto como
soltar el botón, y no lo hacía: las salidas y los ajustes se piden a la
vez, así que la barra podía aparecer con los ajustes todavía en camino y
guardar un archivo que la preferencia decía que no.

Y el encuadre de las salidas suma saturando. Los números los informa el
compositor: una geometría absurda desbordaba, que en depuración es un
pánico —la captura perdida por un dato que no controlamos— en vez de un
rectángulo sin sentido pero finito, que es lo que el resto ya sabe tratar.
El aviso de la cuenta regresiva conserva su identificador si una
notificación falla, para no apilar cinco sobre lo que se va a fotografiar.

+++ src-tauri/src/comandos.rs
@@ -547,8 +547,11 @@ mod pruebas {
         let layout = captura::layout_de(&salidas);
 
         // Cubren todo: la suma de las áreas es la del encuadre.
-        let suma: i64 = salidas.iter().map(|s| i64::from(s.ancho * s.alto)).sum();
-        assert_eq!(suma, i64::from(layout.ancho * layout.alto));
+        let suma: i64 = salidas
+            .iter()
+            .map(|s| i64::from(s.ancho) * i64::from(s.alto))
+            .sum();
+        assert_eq!(suma, i64::from(layout.ancho) * i64::from(layout.alto));
 
         // Y no se pisan: la de abajo empieza donde termina la de arriba.
         assert_eq!(salidas[0].y + salidas[0].alto, salidas[1].y);
